Paper Summary Table
This table distinguishes papers from the original search from papers in
the later search. It also contains information on the journal from which
papers were taken, whether or not a paper was included or excluded,
year, location, and measurements. It is the updated file from the
original manuscript with the new information from my searches.