Modify script to ignore comparisons that fail
Stats script now quietly ignores failure of the escalc() function due to
too little data. Still, check the results csv file to make sure all the
comparisons expected are there.

To accomplish the above, I made a new function that handles a single
comparison. The code is ungly and depends upon eval(parse()) to make
the variable names.  But this now allows wrapping the inidvidual
comparison calls in the tryCatch function.