Completed data column for fire intensity levels in papers.csv
Fire intensity has been pulled from each paper and included in
papers.csv. Fires were categorized as "low" "medium" or "high". My
rational for each assignment is present in a separate file on my
computer lab notebook.

I also removed Freeman et al. from exotic-cover.csv and
exotic-richness.csv, changed it from "used" to "not used" in papers.csv,
and took note of its removal in decisions.csv. It was removed because
the data was not usable (because we needed data from each individual
site, which was not available in the paper).