Moved YearsSinceTreatment to papers.csv, manuscript edits
Moved the YearsSinceTreatment data to papers.csv, to keep everything in
one place.

I also made edits to the manuscript, but I'm waiting to finish it until
my meeting with Dr. Schwilk on 2/18/2014 to finalize the data. One major
edit was in the mean and median YearsSinceTreatment--on my previous data
pull on this I made several errors, in which I over-reported the number
of years since treatment. This has been fixed.